add new utils file to repository package
