+++ internal/repository/repository.go
@@ -1,10 +1,8 @@
 package repository
 
 import (
-	"errors"
 	"expense-api/internal/model"
 
-	"github.com/jackc/pgconn"
 	"gorm.io/gorm"
 )
 
@@ -36,33 +34,10 @@ type Repository interface {
 	TransactionListByParty(userID, partyID uint) ([]*model.Transaction, error)
 }
 
-var (
-	ErrorRecordNotFound           = errors.New("resource not found")
-	ErrorOther                    = errors.New("an error occurred")
-	ErrorUniqueConstaintViolation = errors.New("record already exists (duplicate unique key)")
-)
-
 type repository struct {
 	db *gorm.DB
 }
 
 func New(db *gorm.DB) Repository {
 	return &repository{db}
-}
-
-func checkError(err error) error {
-	if isUniqueConstaintViolationError(err) {
-		return ErrorUniqueConstaintViolation
-	} else if err == gorm.ErrRecordNotFound {
-		return ErrorRecordNotFound
-	}
-	return ErrorOther
-}
-
-func isUniqueConstaintViolationError(err error) bool {
-	uniqueConstraintCode := "23505"
-	if err, ok := err.(*pgconn.PgError); ok {
-		return err.Code == uniqueConstraintCode
-	}
-	return false
 }
